refactor: 💡 rename composables folder to keyboard
