+++ app/src/main/java/com/rickyhu/hushkeyboard/HushKeyboardView.kt
@@ -5,7 +5,7 @@ import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.AbstractComposeView
-import com.rickyhu.hushkeyboard.ui.composables.HushKeyboard
+import com.rickyhu.hushkeyboard.ui.keyboard.HushKeyboard
 import com.rickyhu.hushkeyboard.viewmodel.KeyboardViewModel
 
 class HushKeyboardView(context: Context) : AbstractComposeView(context) {
