+++ app/src/main/java/com/rickyhu/hushkeyboard/ui/keyboard/KeyButton.kt
@@ -1,4 +1,4 @@
-package com.rickyhu.hushkeyboard.ui.composables
+package com.rickyhu.hushkeyboard.ui.keyboard
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
